feat: rename sibscriber collection file
